Change file structure and add update home page
