+++ src/components/Footer.jsx
@@ -16,7 +16,7 @@ const navigation = [
 
 export default function Footer() {
   return (
-    <footer className="fixed bottom-0 w-full bg-white">
+    <footer className="w-full bg-white">
       <div className="mx-auto max-w-7xl py-12 px-6 md:flex md:items-center md:justify-between lg:px-8">
         <div className="flex justify-center space-x-6 md:order-2">
           {navigation.map((item) => (
