+++ src/components/dashboard/ProductListLoader.jsx
@@ -1,6 +1,6 @@
-import { React, useContext, useState, useEffect } from 'react';
+import { React, useContext, useEffect, useState } from 'react';
 import ContentLoader from 'react-content-loader';
-import { AppContext } from './AppContextProvider';
+import { AppContext } from '../AppContextProvider';
 
 const ListLoaderCol4 = () => (
   <div className="hidden lg:block">
