+++ src/pages/index.js
@@ -1,6 +1,6 @@
-import Head from 'next/head';
-
 import { Inter } from '@next/font/google';
+import Head from 'next/head';
+import Layout from '../components/Layout';
 
 const inter = Inter({ subsets: ['latin'] });
 
@@ -14,26 +14,48 @@ export default function Home() {
         <link rel="icon" href="/favicon.ico" />
       </Head>
       <main>
-        <div className="flex gap-2">
-          <a
-            href="/auth/sign-up"
-            className="hover:bg-primary-dark focus:ring-primary-light focus:border-primary-darkest active:bg-primary-darkest flex justify-center rounded-md border border-transparent bg-cyan-800 py-2 px-4 text-sm font-medium text-white transition duration-150 ease-in-out focus:outline-none focus:ring-4"
-          >
-            Sign up
-          </a>
-          <a
-            href="/auth/login"
-            className="hover:bg-primary-dark focus:ring-primary-light focus:border-primary-darkest active:bg-primary-darkest flex justify-center rounded-md border border-transparent bg-cyan-800 py-2 px-4 text-sm font-medium text-white transition duration-150 ease-in-out focus:outline-none focus:ring-4"
-          >
-            Login
-          </a>
-          <a
-            href="/search"
-            className="hover:bg-primary-dark focus:ring-primary-light focus:border-primary-darkest active:bg-primary-darkest flex justify-center rounded-md border border-transparent bg-cyan-800 py-2 px-4 text-sm font-medium text-white transition duration-150 ease-in-out focus:outline-none focus:ring-4"
-          >
-            Search
-          </a>
-        </div>
+        <Layout>
+          <div className="relative isolate overflow-hidden bg-white">
+            <div className="mx-auto max-w-7xl px-6 pb-24 pt-10 sm:pb-32 lg:flex lg:px-8 lg:py-40">
+              <div className="mx-auto max-w-2xl lg:mx-0 lg:max-w-xl lg:flex-shrink-0 lg:pt-8">
+                <div className="mt-24 sm:mt-32 lg:mt-16">
+                  <a href="#" className="inline-flex space-x-6">
+                    <span className="rounded-full bg-black/10 px-3 py-1 text-sm font-semibold leading-6 text-black ring-1 ring-inset ring-black/10">
+                      v1.0 Beta
+                    </span>
+                  </a>
+                </div>
+                <h1 className="mt-10 text-4xl font-bold tracking-tight text-gray-900 sm:text-6xl">
+                  Search any product you want...
+                </h1>
+                <p className="mt-6 text-lg leading-8 text-gray-600">
+                  A product finder for best prices in the market
+                </p>
+                <div className="mt-10 flex items-center gap-x-6">
+                  <a
+                    href="/auth/sign-up"
+                    className="rounded-md bg-[#3bb77e] px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-transparent hover:text-black focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-black"
+                  >
+                    Get started <span aria-hidden="true">→</span>
+                  </a>
+                </div>
+              </div>
+              <div className="mx-auto mt-16 flex max-w-2xl sm:mt-24 lg:ml-10 lg:mr-0 lg:mt-0 lg:max-w-none lg:flex-none xl:ml-32">
+                <div className="max-w-3xl flex-none sm:max-w-5xl lg:max-w-none">
+                  <div className="-m-2 rounded-xl bg-gray-900/5 p-2 ring-1 ring-inset ring-gray-900/10 lg:-m-4 lg:rounded-2xl lg:p-4">
+                    <img
+                      src="/home-placeholder.png"
+                      alt="App screenshot"
+                      width={2432}
+                      height={500}
+                      className="w-[76rem] rounded-md shadow-2xl ring-1 ring-gray-900/10"
+                    />
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+        </Layout>
       </main>
     </>
   );
